Removing this optimized memcpy from library because it is only used for copying to framebuffer, which is done by the kernel.

+++ libustd/ds/ustring.h
@@ -166,7 +166,7 @@ private:
   int   _capacity;
 
 public:
-  static upan::string to_string(uint32_t uiNumber) {
+  static upan::string to_string(uint64_t uiNumber) {
     char strNumber[128];
     unsigned i = 0;
